+++ NEWS.md
@@ -1,6 +1,7 @@
 # naomi 0.0.66
 
 * Use logit offsets for ages below which survey data are available. This is principally used for age 0-14 when survey data start at age 15.
+* Add option `rho_paed_15to49f_ratio` to calculate paediatric prevalence as ratio of female 15-49 adult prevalence.
 
 # naomi 0.0.65